TG-5:fix some details , validations and postman collection

+++ src/controllers/admins.js
@@ -13,14 +13,6 @@ const editAdmin = async (req, res) => {
       });
     }
 
-    if (Object.entries(editAdmin).length === 0) {
-      return res.status(400).json({
-        message: 'Edited admin is empty',
-        data: undefined,
-        error: true,
-      });
-    }
-
     await Admins.findByIdAndUpdate(id, updatedAdmin);
     const result = await Admins.findById(id);
 
@@ -41,10 +33,31 @@ const editAdmin = async (req, res) => {
 const deleteAdmin = async (req, res) => {
   try {
     const { id } = req.params;
+    const result = await Admins.findByIdAndDelete(id);
+
+    if (id === null) {
+      return res.status(400).json({
+        message: 'no id parameter',
+        data: undefined,
+        error: true,
+      });
+    }
 
     await Admins.findByIdAndDelete(id);
 
-    return res.status(204).json();
+    if (result === null) {
+      return res.status(404).json({
+        message: 'Admin not found',
+        data: undefined,
+        error: true,
+      });
+    }
+
+    return res.status(200).json({
+      message: `Admin with id ${id} deleted`,
+      data: result,
+      error: false,
+    });
   } catch (error) {
     return res.status(404).json({
       message: `No admin with '${req.params.id}' as an id`,
